requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
requests==2.32.3
numpy==2.2.1
scipy==1.14.1
soundfile==0.12.1
pydub==0.25.1
python-dotenv==1.0.1
elevenlabs==1.50.3
azure-storage-blob==12.19.0
